# specs/002-todo-v2/contracts/api.yaml
openapi: 3.0.3
info:
  title: TODO App API
  version: 2.0.0
  description: TODO application REST API with date-based management and categories

servers:
  - url: http://localhost:3000/api
    description: Local development server

paths:
  /todos:
    get:
      summary: 특정 날짜의 TODO 항목 조회
      operationId: getTodos
      parameters:
        - name: date
          in: query
          required: false
          description: 조회할 날짜 (YYYY-MM-DD). 미입력 시 오늘 날짜
          schema:
            type: string
            format: date
            example: "2026-01-29"
      responses:
        '200':
          description: TODO 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'

    post:
      summary: 새 TODO 항목 생성
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: TODO 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: 잘못된 요청 (빈 content)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /todos/{id}:
    patch:
      summary: TODO 항목 완료 상태 토글
      operationId: toggleTodo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 토글 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: TODO 항목 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: TODO 항목 삭제
      operationId: deleteTodo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 삭제 성공
        '404':
          description: TODO 항목 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /todos/copy-to-next-day:
    post:
      summary: 미완료 TODO를 다음 날짜로 복사
      operationId: copyToNextDay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyRequest'
      responses:
        '200':
          description: 복사 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CopyResponse'
        '400':
          description: 복사할 미완료 항목 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Category:
      type: string
      enum:
        - work
        - personal
      description: TODO 카테고리 (회사/개인)

    Todo:
      type: object
      required:
        - id
        - content
        - completed
        - category
        - date
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        content:
          type: string
          minLength: 1
          example: "주간 보고서 작성"
        completed:
          type: boolean
          example: false
        category:
          $ref: '#/components/schemas/Category'
        date:
          type: string
          format: date
          example: "2026-01-29"
        createdAt:
          type: string
          format: date-time
          example: "2026-01-29T10:30:00.000Z"

    CreateTodoRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          example: "주간 보고서 작성"
        category:
          $ref: '#/components/schemas/Category'
          default: personal
        date:
          type: string
          format: date
          description: 미입력 시 오늘 날짜
          example: "2026-01-29"

    CopyRequest:
      type: object
      required:
        - sourceDate
      properties:
        sourceDate:
          type: string
          format: date
          description: 복사할 원본 날짜
          example: "2026-01-29"

    CopyResponse:
      type: object
      required:
        - copiedCount
        - targetDate
        - copiedTodos
      properties:
        copiedCount:
          type: integer
          description: 복사된 TODO 개수
          example: 3
        targetDate:
          type: string
          format: date
          description: 복사된 대상 날짜 (원본 + 1일)
          example: "2026-01-30"
        copiedTodos:
          type: array
          items:
            $ref: '#/components/schemas/Todo'

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Content cannot be empty"